Refactor structure project
