+++ lib/ui/add/add_task_screen.dart
@@ -6,7 +6,7 @@ import 'package:hexcolor/hexcolor.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'package:intl/intl.dart';
 import 'package:shared_preferences/shared_preferences.dart';
-import 'package:todo_app/TaskItem.dart';
+import 'package:todo_app/ui/base_component/task_item.dart';
 
 class AddTaskScreen extends StatelessWidget {
   const AddTaskScreen();
